refactor: move service files into dedicated service packages
