+++ src/main/java/com/musfiul/idrrateaggregator/service/resolver/FetchResolver.java
@@ -1,6 +1,7 @@
-package com.musfiul.idrrateaggregator.service;
+package com.musfiul.idrrateaggregator.service.resolver;
 
 import com.musfiul.idrrateaggregator.constant.ResourceType;
+import com.musfiul.idrrateaggregator.service.fetcher.IDRDataFetcher;
 
 public interface FetchResolver {
     IDRDataFetcher resolve(ResourceType type);
