+++ src/test/java/com/musfiul/idrrateaggregator/runner/FinanceStartupRunnerTest.java
@@ -3,9 +3,9 @@ package com.musfiul.idrrateaggregator.runner;
 import com.musfiul.idrrateaggregator.constant.ResourceType;
 import com.musfiul.idrrateaggregator.dto.LatestRatesResponse;
 import com.musfiul.idrrateaggregator.dto.api.APIResponseDTO;
-import com.musfiul.idrrateaggregator.service.FetchResolver;
-import com.musfiul.idrrateaggregator.service.FinanceDataStoreService;
-import com.musfiul.idrrateaggregator.service.IDRDataFetcher;
+import com.musfiul.idrrateaggregator.service.resolver.FetchResolver;
+import com.musfiul.idrrateaggregator.service.data.FinanceDataStoreService;
+import com.musfiul.idrrateaggregator.service.fetcher.IDRDataFetcher;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
